timers: Split out timer_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performace

Cutting down on sched.h dependencies: this is going to be used in
workqueue_types.h in the next patch, so we can kill the sched.h
dependency on workqueue.h.

Conflicts:
	include/linux/timer.h
[add timer_types.h to the code]
(cherry picked from commit dff0fd233a5104337069603d201f8cad74bc0e5a)
